+++ drift/test/integration_tests/regress_2744_test.dart
@@ -1,3 +1,4 @@
+import 'package:async/async.dart';
 import 'package:drift/drift.dart';
 import 'package:test/test.dart';
 
@@ -8,12 +9,12 @@ void main() {
   test('updates after transaction', () async {
     // Regression test for https://github.com/simolus3/drift/issues/2744
     final db = TodoDb(testInMemoryDatabase());
-    final categories = db.categories.all().watch();
+    final categories = StreamQueue(db.categories.all().watch());
 
-    expect(categories, emits(isEmpty));
+    await expectLater(categories, emits(isEmpty));
     await db.categories
         .insertOne(CategoriesCompanion.insert(description: 'desc1'));
-    expect(categories, emits(hasLength(1)));
+    await expectLater(categories, emits(hasLength(1)));
 
     await db.categories.deleteAll();
     await db.batch((batch) {
@@ -21,7 +22,7 @@ void main() {
           db.categories, CategoriesCompanion.insert(description: 'desc2'));
     });
 
-    expect(
+    await expectLater(
         categories,
         emits([
           isA<Category>().having((e) => e.description, 'description', 'desc2')